+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# 0.0.16
+- Fix extension startup with no `workspaceFolder`  
+  you can open a folder then go to a nim file not related to this folder, when you reopen this project, editor document is that file, `getWorkspaceFolder` from that file is `undefined`
+
 # 0.0.15
 - Add check Nim file command